chore: common 폴더에 넣기
